chore: モックデータの削除とスクリプトの整理
- frontend/src/lib/mockData.ts: 削除
- frontend/src/lib/constants.ts: カテゴリ設定を移動
- AssetTrendsSection.tsx: インポートパスを修正
- backend/scripts/: データ操作用スクリプトを移動
