--- frontend/src/lib/mockData.ts
@@ -1,199 +1,0 @@
-export const monthlyDataRaw = [
-  {
-    date: "1月",
-    日本株: 890000,
-    米国株: 1200000,
-    投資信託: 800000,
-    現金: 500000,
-  },
-  {
-    date: "2月",
-    日本株: 920000,
-    米国株: 1150000,
-    投資信託: 820000,
-    現金: 510000,
-  },
-  {
-    date: "3月",
-    日本株: 980000,
-    米国株: 1300000,
-    投資信託: 850000,
-    現金: 520000,
-  },
-  {
-    date: "4月",
-    日本株: 1050000,
-    米国株: 1280000,
-    投資信託: 880000,
-    現金: 530000,
-  },
-  {
-    date: "5月",
-    日本株: 1100000,
-    米国株: 1350000,
-    投資信託: 900000,
-    現金: 540000,
-  },
-  {
-    date: "6月",
-    日本株: 1080000,
-    米国株: 1400000,
-    投資信託: 920000,
-    現金: 550000,
-  },
-  {
-    date: "7月",
-    日本株: 1150000,
-    米国株: 1450000,
-    投資信託: 950000,
-    現金: 560000,
-  },
-  {
-    date: "8月",
-    日本株: 1120000,
-    米国株: 1380000,
-    投資信託: 970000,
-    現金: 570000,
-  },
-  {
-    date: "9月",
-    日本株: 1180000,
-    米国株: 1420000,
-    投資信託: 990000,
-    現金: 580000,
-  },
-  {
-    date: "10月",
-    日本株: 1220000,
-    米国株: 1500000,
-    投資信託: 1010000,
-    現金: 590000,
-  },
-  {
-    date: "11月",
-    日本株: 1280000,
-    米国株: 1550000,
-    投資信託: 1030000,
-    現金: 600000,
-  },
-  {
-    date: "12月",
-    日本株: 1350000,
-    米国株: 1600000,
-    投資信託: 1050000,
-    現金: 610000,
-  },
-];
-
-export const yearlyDataRaw = [
-  {
-    date: "2020",
-    日本株: 650000,
-    米国株: 800000,
-    投資信託: 500000,
-    現金: 400000,
-  },
-  {
-    date: "2021",
-    日本株: 780000,
-    米国株: 950000,
-    投資信託: 620000,
-    現金: 450000,
-  },
-  {
-    date: "2022",
-    日本株: 850000,
-    米国株: 1100000,
-    投資信託: 720000,
-    現金: 480000,
-  },
-  {
-    date: "2023",
-    日本株: 1100000,
-    米国株: 1350000,
-    投資信託: 880000,
-    現金: 550000,
-  },
-  {
-    date: "2024",
-    日本株: 1350000,
-    米国株: 1600000,
-    投資信託: 1050000,
-    現金: 610000,
-  },
-];
-
-export const dailyDataRaw = [
-  {
-    date: "12/1",
-    日本株: 1300000,
-    米国株: 1550000,
-    投資信託: 1020000,
-    現金: 600000,
-  },
-  {
-    date: "12/5",
-    日本株: 1310000,
-    米国株: 1560000,
-    投資信託: 1025000,
-    現金: 602000,
-  },
-  {
-    date: "12/10",
-    日本株: 1320000,
-    米国株: 1555000,
-    投資信託: 1030000,
-    現金: 604000,
-  },
-  {
-    date: "12/15",
-    日本株: 1335000,
-    米国株: 1580000,
-    投資信託: 1040000,
-    現金: 606000,
-  },
-  {
-    date: "12/20",
-    日本株: 1340000,
-    米国株: 1590000,
-    投資信託: 1045000,
-    現金: 608000,
-  },
-  {
-    date: "12/25",
-    日本株: 1350000,
-    米国株: 1600000,
-    投資信託: 1050000,
-    現金: 610000,
-  },
-];
-
-export const categoryConfig = {
-  合計: {
-    color: "cyan",
-    bgColor: "bg-cyan-500",
-    borderColor: "border-cyan-500",
-  },
-  日本株: {
-    color: "indigo",
-    bgColor: "bg-indigo-500",
-    borderColor: "border-indigo-500",
-  },
-  米国株: {
-    color: "amber",
-    bgColor: "bg-amber-500",
-    borderColor: "border-amber-500",
-  },
-  投資信託: {
-    color: "emerald",
-    bgColor: "bg-emerald-500",
-    borderColor: "border-emerald-500",
-  },
-  現金: {
-    color: "slate",
-    bgColor: "bg-slate-500",
-    borderColor: "border-slate-500",
-  },
-} as const;
-
-export type CategoryKey = keyof typeof categoryConfig;
